Refactor google cloud storage class
